refactoered the Unit struct

+++ appserver/src/objects/units/class/warrior.h
@@ -1,16 +1,17 @@
 #pragma once
+#include <cstdint>
 
 namespace Warrior {
 
     struct Stats {
-        int x = 6;
-        int y = 6;
-        int health = 100;
-        int healthMax = 100;
-        int attack = 10;
-        int defense = 5;
-        int speed = 5;
-        int vision = 6;
+        uint8_t x = 6;
+        uint8_t y = 6;
+        uint8_t health = 100;
+        uint8_t healthMax = 100;
+        uint8_t attack = 10;
+        uint8_t defense = 5;
+        uint8_t speed = 5;
+        uint8_t vision = 6;
     };
 
     //Stats Player;
